consolidate all page-related components to their own directories
