+++ ohmg/frontend/svelte_components/src/map.js
@@ -1,4 +1,4 @@
-import Map from './components/Map.svelte';
+import Map from './components/map/Map.svelte';
 
 export default new Map({
   target: document.getElementById('map-target'),
