+++ ohmg/frontend/svelte_components/src/resource.js
@@ -1,4 +1,4 @@
-import Resource from './components/Resource.svelte';
+import Resource from './components/resource/Resource.svelte';
 
 export default new Resource({
   target: document.getElementById('resource-target'),
